perception/requirements.txt
open3d~=0.19.0
numpy~=2.2.3

opencv-python~=4.11.0.86
mcap~=1.2.2
mcap-ros2-support==0.5.5

boto3~=1.36.26
botocore~=1.36.26
PyYAML~=6.0.2
pytz~=2025.1

pydantic~=2.10.6
pydantic-settings~=2.8.0

requests~=2.32.3

pandas~=2.2.3